[JIT] Support for ParameterDict getattr
Adds support for scripting ParameterDicts and getattr() on them. It does
not support iterating on ParameterDicts because torch/nn/container.py
implementation of ParameterDict.items() uses a generator, which is not
supported by torchscript. torch/nn/container.py would need to be updated
so that iter gets correctly registered in python_sugared_value.cpp

Added a test in test_module_containers.py

ghstack-source-id: 01e36adaf24526006e6f9ba839a03eaa8323fc70
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77143